Add SerializerDecorator that will remove non-sendable stamps and then call the decorated serializer.
