+++ src/Symfony/Component/Messenger/Tests/Transport/Serialization/PhpSerializerTest.php
@@ -77,18 +77,6 @@ class PhpSerializerTest extends TestCase
         ]);
     }
 
-    public function testEncodedSkipsNonEncodeableStamps()
-    {
-        $serializer = new PhpSerializer();
-
-        $envelope = new Envelope(new DummyMessage('Hello'), [
-            new DummyPhpSerializerNonSendableStamp(),
-        ]);
-
-        $encoded = $serializer->encode($envelope);
-        $this->assertStringNotContainsString('DummyPhpSerializerNonSendableStamp', $encoded['body']);
-    }
-
     public function testNonUtf8IsBase64Encoded()
     {
         $serializer = new PhpSerializer();
@@ -99,8 +87,4 @@ class PhpSerializerTest extends TestCase
         $this->assertTrue((bool) preg_match('//u', $encoded['body']), 'Encodes non-UTF8 payloads');
         $this->assertEquals($envelope, $serializer->decode($encoded));
     }
-}
-
-class DummyPhpSerializerNonSendableStamp implements NonSendableStampInterface
-{
 }
