+++ src/Symfony/Component/Messenger/Transport/Serialization/PhpSerializer.php
@@ -43,8 +43,6 @@ class PhpSerializer implements SerializerInterface
      */
     public function encode(Envelope $envelope): array
     {
-        $envelope = $envelope->withoutStampsOfType(NonSendableStampInterface::class);
-
         $body = addslashes(serialize($envelope));
 
         if (!preg_match('//u', $body)) {
