+++ src/Symfony/Component/Messenger/Transport/Serialization/SerializerDecorator.php
@@ -1,0 +1,44 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Messenger\Transport\Serialization;
+
+use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\Exception\MessageDecodingFailedException;
+use Symfony\Component\Messenger\Stamp\NonSendableStampInterface;
+
+class SerializerDecorator implements SerializerInterface
+{
+    private $serializer;
+
+    public function __construct(SerializerInterface $serializer)
+    {
+        $this->serializer = $serializer;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function decode(array $encodedEnvelope): Envelope
+    {
+        return $this->serializer->decode($encodedEnvelope);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function encode(Envelope $envelope): array
+    {
+        $envelope = $envelope->withoutStampsOfType(NonSendableStampInterface::class);
+
+        return $this->serializer->encode($envelope);
+    }
+}
